# openapi_contract.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Servidor App Entradas",
    "version": "1.0.0",
    "description": "API para gestión de entradas y salidas de vehículos, personal y empresas externas",
    "contact": {
      "name": "Soporte",
      "email": "soporte@example.com"
    }
  },
  "servers": [
    {
      "url": "http://10.192.93.0:7302",
      "description": "Servidor de Desarrollo (10.192.93.0)"
    },
    {
      "url": "http://10.192.92.12:7202",
      "description": "Servidor de Producción (10.192.92.12)"
    },
    {
      "url": "http://localhost:7302",
      "description": "Localhost (si estás en el servidor)"
    }
  ],
  "components": {
    "securitySchemes": {
      "xTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-token",
        "description": "JWT emitido por /api/login en header x-token"
      }
    },
    "schemas": {
...
      "EntradaVehiculo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "empresa": {
            "type": "string",
            "example": "Transportes SA"
          },
          "nombre_conductor": {
            "type": "string",
            "example": "Pedro García"
          },
          "matricula": {
            "type": "string",
            "example": "ABC-1234"
          },
          "clase_carga": {
            "type": "string",
            "example": "Material de construcción"
          },
          "recepcion": {
            "type": "boolean",
            "example": false
          },
          "vigilancia": {
            "type": "boolean",
            "example": false
          },
          "firma": {
            "type": "string",
            "description": "Firma del conductor en formato base64",
            "example": "data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA..."
          },
          "fecha_entrada": {
            "type": "string",
            "format": "date-time"
          },
          "fecha_salida": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "fecha_creacion": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación del registro"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "xTokenAuth": []
    }
  ],
  "paths": {
    "/api/entradas": {
      "get": {
        "summary": "Obtener vehículos dentro (últimas 12h o sin salida)",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de entradas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntradaVehiculo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear nueva entrada de vehículo",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "empresa",
                  "matricula",
                  "fecha_entrada"
                ],
                "properties": {
                  "empresa": {
                    "type": "string"
                  },
                  "nombre_conductor": {
                    "type": "string"
                  },
                  "matricula": {
                    "type": "string"
                  },
                  "clase_carga": {
                    "type": "string"
                  },
                  "fecha_entrada": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "firma": {
                    "type": "string",
                    "description": "Firma del conductor en formato base64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entrada creada"
          },
          "400": {
            "description": "Datos inválidos"
          }
        }
      }
    },
    "/api/entradas/almacen": {
      "get": {
        "summary": "Obtener entradas para almacén",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entradas para almacén"
          }
        }
      }
    },
    "/api/entradas/porteria": {
      "get": {
        "summary": "Obtener entradas para portería",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Entradas para portería"
          }
        }
      },
      "put": {
        "summary": "Actualizar entrada en portería",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "fecha": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "vigilancia": {
                    "type": "boolean",
                    "description": "Estado de vigilancia (true para marcar salida)"
                  }
                },
                "required": [
                  "id",
                  "vigilancia",
                  "fecha"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrada actualizada"
          }
        }
      }
    },
    "/api/entradas/by-matricula/{matricula}": {
      "get": {
        "summary": "Buscar entrada por matrícula",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "matricula",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Matrícula del vehículo"
          }
        ],
        "responses": {
          "200": {
            "description": "Entrada encontrada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      }
    },
    "/api/entradas/{id}": {
      "get": {
        "summary": "Obtener entrada específica por ID",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la entrada"
          }
        ],
        "responses": {
          "200": {
            "description": "Entrada encontrada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "delete": {
        "summary": "Eliminar entrada",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entrada eliminada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      },
      "put": {
        "summary": "Actualizar entrada completa",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "empresa": {
                    "type": "string"
                  },
                  "nombre_conductor": {
                    "type": "string"
                  },
                  "matricula": {
                    "type": "string"
                  },
                  "clase_carga": {
                    "type": "string"
                  },
                  "fecha_entrada": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "fecha_salida": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "firma": {
                    "type": "string",
                    "description": "Firma del conductor en formato base64"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entrada actualizada"
          },
          "404": {
            "description": "No encontrada"
          }
        }
      }
    },
    "/api/entradas/recepcion": {
      "put": {
        "summary": "Actualizar estado de recepción",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "integer"
                  },
                  "recepcion": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          }
        }
      }
    },
    "/api/entradas/select": {
      "put": {
        "summary": "Consultar entradas por rango de fechas",
        "tags": [
          "Entradas"
        ],
        "security": [
          {
            "xTokenAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fecha_entrada1",
                  "fecha_entrada2"
                ],
                "properties": {
                  "fecha_entrada1": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "fecha_entrada2": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Lista filtrada de entradas"
          }
        }
      }
    },
...
  },
...
}